Fix for rerunning events

+++ backend/app/api/platform/models/events.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 
 class EventBackfillRequest(BaseModel):
-    created_at_start: Optional[int] = None
-    created_at_end: Optional[int] = None
+    created_at_start: Optional[float] = None
+    created_at_end: Optional[float] = None
     event_id: str
     sample_rate: float = Field(ge=0, le=1)
